Share mapping test between clean-up commands
