+++ datahub/cleanup/test/commands/test_common.py
@@ -12,7 +12,15 @@ from django.utils.timezone import utc
 from freezegun import freeze_time
 
 from datahub.cleanup.management.commands import delete_old_records, delete_orphans
-from datahub.cleanup.test.commands.test_delete_orphans import MAPPINGS
+from datahub.cleanup.query_utils import get_related_fields
+from datahub.cleanup.test.commands.factories import ShallowInvestmentProjectFactory
+from datahub.company.test.factories import CompanyFactory, ContactFactory
+from datahub.event.test.factories import EventFactory
+from datahub.interaction.test.factories import CompanyInteractionFactory
+from datahub.investment.test.factories import InvestmentProjectFactory
+from datahub.leads.test.factories import BusinessLeadFactory
+from datahub.omis.order.test.factories import OrderFactory
+from datahub.omis.quote.test.factories import QuoteFactory
 from datahub.search.apps import get_search_app_by_model, get_search_apps
 
 COMMAND_CLASSES = [
@@ -21,6 +29,50 @@ COMMAND_CLASSES = [
 ]
 
 FROZEN_TIME = datetime(2018, 6, 1, 2, tzinfo=utc)
+
+"""
+For each model in the CONFIG attribute for clean-up management commands:
+    - specify the factory class to create an instance
+    - specify the list of dependent models as a tuple of:
+        - dependent factory class to create an instance
+        - dependent field name referencing the main model
+"""
+MAPPINGS = {
+    'company.Contact': {
+        'factory': ContactFactory,
+        'dependent_models': (
+            (CompanyInteractionFactory, 'contact'),
+            (OrderFactory, 'contact'),
+            (QuoteFactory, 'accepted_by'),
+            (InvestmentProjectFactory, 'client_contacts'),
+        )
+    },
+    'company.Company': {
+        'factory': CompanyFactory,
+        'dependent_models': (
+            (CompanyInteractionFactory, 'company'),
+            (ContactFactory, 'company'),
+            (ShallowInvestmentProjectFactory, 'intermediate_company'),
+            (ShallowInvestmentProjectFactory, 'investor_company'),
+            (ShallowInvestmentProjectFactory, 'uk_company'),
+            (OrderFactory, 'company'),
+            (CompanyFactory, 'global_headquarters'),
+            (CompanyFactory, 'parent'),
+            (BusinessLeadFactory, 'company'),
+        )
+    },
+    'event.Event': {
+        'factory': EventFactory,
+        'dependent_models': (
+            (CompanyInteractionFactory, 'event'),
+        )
+
+    },
+    'interaction.Interaction': {
+        'factory': CompanyInteractionFactory,
+        'dependent_models': (),
+    }
+}
 
 
 def _format_iterable(value):
@@ -68,6 +120,11 @@ def cleanup_args(request):
     yield (request.param[0](), *request.param[1:])
 
 
+def _format_cleanup_mapping(cleanup_mapping):
+    command, model_name, config, mapping, dep_factory, dep_field_name = cleanup_mapping
+    return _format_iterable((command, model_name, dep_field_name))
+
+
 @pytest.fixture(params=(
     (
         command_cls,
@@ -80,11 +137,48 @@ def cleanup_args(request):
     for command_cls in COMMAND_CLASSES
     for model_name, config in command_cls.CONFIGS.items()
     for dep_factory, dep_field_name in MAPPINGS[model_name]['dependent_models']
-), ids=_format_iterable)
+), ids=_format_cleanup_mapping
+)
 def cleanup_mapping(request):
-    """Fixture that parametrises tests for each clean-up command configuration."""
+    """Fixture that parametrises tests for each clean-up command, model and relationship field."""
     # Instantiate the command class
     yield (request.param[0](), *request.param[1:])
+
+
+@pytest.mark.parametrize(
+    'model_name',
+    # Get unique models only – must be in a consistent order for parallelised tests
+    sorted(
+        {model_name for command_cls in COMMAND_CLASSES for model_name in command_cls.CONFIGS}
+    )
+)
+def test_mappings(model_name):
+    """
+    Test that `MAPPINGS` includes all the data necessary for covering all the cases.
+    This is to avoid missing tests when new fields and models are added or changed.
+    """
+    model = apps.get_model(model_name)
+
+    try:
+        mapping = MAPPINGS[model_name]
+    except KeyError:
+        pytest.fail(f'Please add test cases for deleting orphaned {model}')
+
+    related_fields = get_related_fields(model)
+    expected_related_deps = {(field.field.model, field.field.name) for field in related_fields}
+    related_deps_in_mapping = {
+        (dep_factory._meta.model, dep_field_name)
+        for dep_factory, dep_field_name in mapping['dependent_models']
+    }
+
+    missing_dep_mappings = expected_related_deps - related_deps_in_mapping
+    if missing_dep_mappings:
+        dep_list = [f'{model}.{field}' for model, field in missing_dep_mappings]
+        error_msg = (
+            f'Please add tests for not deleting {model} when the following '
+            f'fields reference it: {", ".join(dep_list)}'
+        )
+        assert not missing_dep_mappings, error_msg
 
 
 def create_orphanable_model(factory, config, date_value):
